+++ CHANGELOG.md
@@ -8,6 +8,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 For a full diff see [`2.9.0...main`][2.9.0...main].
 
+### Changed
+
+* Added `Config\RuleSet\ExplicitRuleSet` marker interface for rule-sets that should  be configured explicitly ([#311]), by [@localheinz]
+
 ## [`2.9.0`][2.9.0]
 
 For a full diff see [`2.8.0...2.9.0`][2.8.0...2.9.0].
@@ -294,6 +298,7 @@ For a full diff see [`d899e77...1.0.0`][d899e77...1.0.0].
 [#306]: https://github.com/ergebnis/php-cs-fixer-config/pull/306
 [#309]: https://github.com/ergebnis/php-cs-fixer-config/pull/309
 [#310]: https://github.com/ergebnis/php-cs-fixer-config/pull/310
+[#311]: https://github.com/ergebnis/php-cs-fixer-config/pull/311
 
 [@dependabot]: https://github.com/apps/dependabot
 [@linuxjuggler]: https://github.com/linuxjuggler